New folder structure + comments :memo:
