+++ Code/1_Ontology_projection/Mesh_to_graph.py
@@ -1,17 +1,15 @@
-# 
-# Install mowl library
-# pip install mowl-borg
-
-
 import sys
 sys.path.append('../../')
 import mowl
 import pickle
 import os
-mowl.init_jvm("10g")
+mowl.init_jvm("4g")
 from mowl.datasets import PathDataset
 from mowl.projection import DL2VecProjector
 from mowl.projection import OWL2VecStarProjector
+
+# Data: mesh ontologies from https://bioportal.bioontology.org/ontologies/MESH?p=summary
+dataset = PathDataset('Data/Input/MESH.ttl')
 
 # Setting parameters of the OWL2VEC projector
 projector = OWL2VecStarProjector(
@@ -19,11 +17,7 @@ projector = OWL2VecStarProjector(
              include_literals = False,
              only_taxonomy = True)
 
-# Data: mesh ontologies from https://bioportal.bioontology.org/ontologies/MESH?p=summary
-dataset = PathDataset('./MESH.ttl')
-
-# Project the ontologies to vector
-projector = DL2VecProjector(bidirectional_taxonomy=True)
+# Project the ontologies to a graph
 edges = projector.project(dataset.ontology)
 
 # Save the graph in a pickle variable
