+++ Code/3_Abstract2Graph/Readme.md
@@ -1,0 +1,5 @@
+## Objective
+
+Experiment with different approaches to create graphs from abstracts using the MeSH ontology concepts present in the abstract.
+
+insert image:
