+++ Code/4_Graph_Enrichment/Readme.md
@@ -1,0 +1,9 @@
+## Objective
+Here we are creating new graphs by doing traversals through the MeSH ontology projected graph.
+
+The main strategy from v1 to v4 consists of finding the smallest path between 2 nodes that have an edge 
+in the initial graph that represents an abstract, we collect all the unique nodes that appear in all the path traversals and that  basically represents our new graph for an abstract.
+
+The difference between the versions  v1 to v4 basically are performance improvements. 
+
+insert image:
