+++ Code/5_Abstract_Enrichment/Abstract_enrichment_v1.py
@@ -1,0 +1,31 @@
+"""
+    Here we want to add more text to the RELISH abstracts. The new text will be generated from an ontology, and will
+    correspond to the paths that can be generated by traversing the ontology using the concepts found in the abstract.
+"""
+from Bio import Entrez
+import timeit
+import json
+import pickle
+import mowl
+mowl.init_jvm("4g")
+import networkx as nx
+
+
+def main():
+    Abstract_Graph_raw = []
+    with open('../../Data/Output/file_single_mesh.json') as f:
+        Abstract_Graph_raw = json.load(f)
+
+    Onto_graph_raw = []
+    with (open("../../Data/Input/Graph_mesh_OWL2VEC.pkl", "rb")) as openfile:
+        while True:
+            try:
+                Onto_graph_raw.append(pickle.load(openfile))
+            except EOFError:
+                break    
+
+t_0 = timeit.default_timer()
+main()
+t_1 = timeit.default_timer()
+elapsed_time = round((t_1 - t_0), 3)
+print(f"Elapsed time: {elapsed_time}")
